Made API to let CircuitVerse save it's creation in two formats: A) CircuitVerse readable Json file, and B) ControlPi readable conf file. (Now via tools -> export Logic )
"Save Online" in CircuitVerse now makes a  ajax call to logicUpdateApi.php and saves to  of logic.json
On startup, the same PHP Script is called, which reads logic.json and gives its contents back to Cirquitverse.
The conf file, which is really is responsible for the PLC control must not ne changed manually, so logic.json and logic.conf won't get out of sync.

	modified:   www/logic-editor/index.html
	modified:   www/logic-editor/js/UX.js
	modified:   www/logic-editor/js/data.js
	modified:   www/logic-editor/js/logicExport.js
	modified:   www/logic-editor/js/logix.js
	modified:   www/logic.conf
	modified:   www/logic.json
	modified:   www/logicUpdateApi.php